% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{bradtheta}
\alias{bradtheta}
\title{Calculates Bradford's hydrotime constant (Theta).}
\usage{
bradtheta(probit, x, invt)
}
\arguments{
\item{probit}{a vector of probit transformations of the cumulative germination proportion}

\item{x}{a vector of water potentials of the experimental treatment}

\item{invt}{a vector of the inverse of the germination scoring time since the start of
the experiment}
}
\value{
\code{bradtheta} returns the value of Theta and passes it back to \code{\link{bradford}}
}
\description{
\code{bradtheta} calculates the population hydrotime constant (Theta)
  using Bradford's method by solving EXPLICAR LOS CALCULOS de GIL. This
  function is mainly used by \code{\link{bradford}} to perform its
  calculations.
}
