% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedr.R
\docType{data}
\name{centaury}
\alias{centaury}
\title{Temperature example dataset}
\format{
A data frame with 896 rows and 7 variables \describe{
  \item{species}{Name or code for the species to which the data refers}
  \item{population}{Name or code for the seedlot}
  \item{temperature}{Temperature treatment (in ºC) of the experiment}
  \item{dish}{Code for the Petri dish, container, replicate, etc.}
  \item{times}{Time of germination scoring since the start of the experiment,
  in days} \item{germinated}{Number of germinated seeds recorded at that
  time} \item{germinable}{Total number of viable seeds in the replicate} }
}
\source{
Own data from a laboratory experiment.
}
\usage{
centaury
}
\description{
This is a dataset containing information from a germination temperature experiment
with centaury seeds. It is used to give examples of the functions dealing with
thermal time germination models. It also gives and idea of the format in which
germination data should be provided to seedr.
}
\keyword{datasets}
