% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedr.R
\docType{package}
\name{seedr}
\alias{seedr}
\title{seedr: A package to fit hydro and thermal time seed germination models}
\description{
The seedr package provides functions to process data generated in laboratory
germination experiments, fitting hydrothermal time germination models. These
models allow to characterize seed lots by parameters that indicate the physiological
thresholds (of temperature or water) between which the seed lot can germinate,
and the physiological-time units that the seeds in the seed lot need to accumulate
before they can germinate.
}
