% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mt.R
\name{test_mt}
\alias{test_mt}
\title{Tests the correlation using Spearman's rank between samples using different
Mersenne-Twister seed setups.}
\usage{
test_mt(num_samples, reps)
}
\arguments{
\item{num_samples}{stub}

\item{reps}{stub}
}
\value{
List of correlation for different seed setups
}
\description{
Tests the correlation using Spearman's rank between samples using different
Mersenne-Twister seed setups.
}
\examples{
\dontrun{
test_mt(num_samples = 10, reps = 100)
}
}
