% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_seed.R
\name{test_seed}
\alias{test_seed}
\title{Tests the correlation using Spearman's rank between samples using differnt
seed setups.}
\usage{
test_seed(num_samples, reps)
}
\arguments{
\item{num_samples}{stub}

\item{reps}{stub}
}
\value{
List of correlation for different seed setups
}
\description{
Tests the correlation using Spearman's rank between samples using differnt
seed setups.
}
\examples{
\dontrun{
test_seed(10, 100)
}
}
