% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wh.R
\name{test_wh}
\alias{test_wh}
\title{Tests the correlation using Spearman's rank between samples using different
Wichmanns-Hill seed setups.}
\usage{
test_wh(num_samples, reps)
}
\arguments{
\item{num_samples}{stub}

\item{reps}{stub}
}
\value{
List of correlation for different seed setups
}
\description{
Tests the correlation using Spearman's rank between samples using different
Wichmanns-Hill seed setups.
}
\examples{
\dontrun{
test_wh(num_samples = 10, reps = 100)
}
}
