% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradford.R
\name{bradford}
\alias{bradford}
\title{Fitting Bradford's hydrotime model.}
\usage{
bradford(d, t, g, pg, x, reps = NULL, groups = NULL)
}
\arguments{
\item{d}{a data.table produced by \code{physiotime} containing the
germination data}
}
\value{
\code{bradford} returns a list with the results which is passed back
  to \code{\link{physiotime}}. The list includes the following elements:
  \describe{ \item{groups}{a data.table with the grouping variables defined
  by the \code{groups} argument} of \code{physiotime} \item{data}{a
  data.table with the original germination data plus new columns with
  variables calculated for model fitting} \item{models}{the \code{lm} objects
  fitted to calculate the physiological time parameters} \item{parameters}{a
  list with a summary of the experiment and the physiological time
  parameters}}
}
\description{
\code{bradford} is used by \code{\link{physiotime}} when \code{method =
"bradford"} to fit a hydrotime seed germination model using the method of
Bradford.
}
