% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiotime.R
\name{physiotime}
\alias{physiotime}
\title{Fitting Physiological Time Seed Germination Models.}
\usage{
physiotime(d, t, g, pg, x, reps = NULL, groups = NULL,
  method)
}
\arguments{
\item{d}{a data frame containing the results of a germination experiment.
The data frame should include columns with scoring times, germination
counts (not cumulative), number of potentially germinable seeds, and the
environmental variable of interest. (e.g. temperature or water potential)
(see \code{\link{peg}} example dataset for appropiate structure)}

\item{t}{the name of a column (quoted) in \code{d} containing a vector of
numeric scoring times}

\item{g}{the name of a column (quoted) in \code{d} containing a vector of
integer germination counts (non cumulative)}

\item{pg}{the name of a column (quoted) in \code{d} containing a vector of
integer numbers of potentially germinable seeds}

\item{x}{the name of a column (quoted) in \code{d} containing a vector of
numeric values for the environmental variable of interest (e.g.
temperature, water potential)}

\item{groups}{optional, the names of columns (quoted) in \code{d} containing
grouping variables for the experiment that have to be analysed separately
(e.g. different species or populations, different tempperatures in a water
potential experiment, different treatments to break seed dormancy)}

\item{method}{the method to be used to fit the models, e.g. Bradford's
hydrotime model}

\item{reps}{optional, the name of a column (quoted) in \code{d} containing a
vector of replicate information (e.g. Petri dish)}
}
\value{
\code{physiotime} returns a S3 object of class equal to the method
  specified by the \code{method} argument. The object is a list. The elements
  in the first level are the experimental groups as specified by the
  \code{groups} argument (if \code{groups = NULL}, there will be a single
  first-level element). The second level of the list contains, for each
  group, the following elements:
  \describe{
  \item{groups}{a data.table with
  the grouping variables defined by the \code{groups} argument}
  \item{data}{a data.table with the original germination data plus new columns with
  variables calculated for model fitting}
  \item{models}{the model objects
  fitted to calculate the physiological time parameters}
  \item{parameters}{a list with a summary of the experiment and the physiological time
  parameters}}
}
\description{
\code{physiotime} fits physiological time models (thermal time, hydrotime) to
seed germination data.
}
\details{
This function allows to specify the method to be used (e.g. Bradford's
hydrotime model). \code{physiotime} is a constructor function that will
return a S3 object whose class depends on the method defined to fit the
models.
}
\examples{
physiotime(d = peg, t = "times", g = "germinated",
                     pg = "germinable", x = "psi", reps = "dish",
                     groups = c("species", "temperature"),
                     method = "bradford")
}
